overhaul, simplify, and combine parameterized linear and quadratic repn visitors
