+++ pyomo/repn/__init__.py
@@ -11,3 +11,22 @@
 
 from pyomo.repn.standard_repn import StandardRepn, generate_standard_repn
 from pyomo.repn.standard_aux import compute_standard_repn
+
+from pyomo.common.deprecation import moved_module
+
+moved_module(
+    "pyomo.repn.parameterized_linear",
+    "pyomo.repn.parameterized",
+    msg="The pyomo.repn.parameterized_linear module is deprecated.  "
+    "Import the ParameterizedLinearRepnVisitor from pyomo.repn.parameterized",
+    version='6.9.3.dev0',
+)
+moved_module(
+    "pyomo.repn.parameterized_quadratic",
+    "pyomo.repn.parameterized",
+    msg="The pyomo.repn.parameterized_quadratic module is deprecated.  "
+    "Import the ParameterizedQuadraticRepnVisitor from pyomo.repn.parameterized",
+    version='6.9.3.dev0',
+)
+
+del moved_module
